backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
python-jose
passlib
alembic
python-multipart